refactor: move implementation into src/ (backward compatible)
- google-home-notifier.js -> src/index.js, .d.ts -> src/index.d.ts (git mv, history preserved)
- package.json main/types now point at src/; files includes src/
- Kept thin root google-home-notifier.{js,d.ts} shims that re-export src, so the
  legacy require('google-home-notifier/google-home-notifier.js') path still works.
- example.js + tests updated to the new path.

require('google-home-notifier') (via main) and the legacy deep path both resolve
to the same module. 9/9 tests green.
